Extract stats methods into `Meilisearch::Index::Stats` module
